gettype fix for drupal 6
git-svn-id: svn://svn.caucho.com/resin/trunk@3926 9c94448d-38f1-0310-a231-d98308ff1ebf

+++ modules/quercus/src/com/caucho/quercus/env/DoubleValue.java
@@ -71,7 +71,9 @@ public class DoubleValue extends NumberValue
    */
   public String getType()
   {
-    return "float";
+    // php/0142
+    
+    return "double";
   }
 
   /**
